feat: add new TypeScript utility functions and documentation
- Introduced new documentation files for variable and function-related APIs, enhancing the TypeScript code generation capabilities.
- Added `genClass`, `genConstructor`, `genClassProperty`, and `genClassMethod` functions for generating TypeScript class structures.
- Implemented control flow utilities including `genIf`, `genElseIf`, `genElse`, `genFor`, `genWhile`, and `genSwitch` for improved code generation.
- Created `genDecorator`, `genModule`, and `genNamespace` functions to support TypeScript module and namespace declarations.
- Updated the SQLite database for documentation content to reflect the new additions and changes.
