+++ src/condition.ts
