+++ src/index.ts
@@ -1,6 +1,21 @@
+export * from "./types";
+export * from "./variable";
+export * from "./type-alias";
+export * from "./conditional";
+export * from "./interface";
+export * from "./enum";
+export * from "./function";
+export * from "./class";
+export * from "./decorator";
+export * from "./module";
+export * from "./namespace";
+export * from "./condition";
+export * from "./try";
+export * from "./loop";
+export * from "./switch";
+export * from "./statement";
 export * from "./esm";
 export * from "./object";
 export * from "./string";
 export * from "./types";
-export * from "./typescript";
 export * from "./utils";
