+++ src/interface.ts
@@ -1,4 +1,4 @@
-import { genJSDocComment, wrapInDelimiters } from "../utils";
+import { genJSDocComment, wrapInDelimiters } from "./utils";
 import { genParam } from "./function";
 import { genTypeObject, genProperty } from "./type-alias";
 import type {
