--- src/typescript/index.ts
@@ -1,97 +1,0 @@
-export type {
-  TypeGeneric,
-  TypeField,
-  FunctionOpts,
-  TypeObjectWithJSDoc,
-  TypeObject,
-  TypeObjectField,
-  GenInterfaceOptions,
-  EnumMemberValue,
-  GenEnumOptions,
-  GenTypeAliasOptions,
-  GenVariableOptions,
-  GenIfOptions,
-  GenElseOptions,
-  GenTryOptions,
-  GenCatchOptions,
-  GenFinallyOptions,
-  GenSwitchOptions,
-  GenClassOptions,
-  GenConstructorOptions,
-  GenClassPropertyOptions,
-  GenClassMethodOptions,
-  GenGetterOptions,
-  GenSetterOptions,
-  GenArrowFunctionOptions,
-  GenMethodOptions,
-  GenCallSignatureOptions,
-  GenConstructSignatureOptions,
-} from "./types";
-
-export { genVariable } from "./variable";
-
-export {
-  genTypeAlias,
-  genTypeExport,
-  genInlineTypeImport,
-  genTypeObject,
-  genProperty,
-  genUnion,
-  genIntersection,
-  genMappedType,
-  genTemplateLiteralType,
-  genKeyOf,
-  genTypeof,
-  genTypeAssertion,
-  genSatisfies,
-} from "./type-alias";
-
-export { genConditionalType } from "./conditional";
-
-export {
-  genInterface,
-  genIndexSignature,
-  genCallSignature,
-  genConstructSignature,
-} from "./interface";
-
-export { genEnum, genConstEnum } from "./enum";
-
-export {
-  genParam,
-  genFunction,
-  genBlock,
-  genArrowFunction,
-  genMethod,
-} from "./function";
-
-export {
-  genClass,
-  genConstructor,
-  genClassProperty,
-  genClassMethod,
-  genGetter,
-  genSetter,
-} from "./class";
-
-export { genDecorator } from "./decorator";
-
-export { genAugmentation, genModule } from "./module";
-
-export { genDeclareNamespace, genNamespace } from "./namespace";
-
-export {
-  genPrefixedBlock,
-  genIf,
-  genElseIf,
-  genElse,
-  genTernary,
-} from "./condition";
-
-export { genTry, genCatch, genFinally } from "./try";
-
-export { genFor, genForOf, genForIn, genWhile, genDoWhile } from "./loop";
-
-export { genCase, genDefault, genSwitch } from "./switch";
-
-export { genThrow, genReturn } from "./statement";
