+++ src/variable.ts
